backend/requirements.txt
# ============================================================
# FINGERPRINT ATTENDANCE SYSTEM - DEPENDENCIES
# ============================================================
# Install all dependencies with: pip install -r requirements.txt
# ============================================================

# Django Web Framework (version 5.2 or higher)
# Core framework for the web application
Django>=5.2

# Pandas - Data Analysis Library
# Used for processing attendance data and creating DataFrames
pandas>=2.2.3

# OpenPyXL - Excel File Handler
# Required for exporting attendance reports to .xlsx format
openpyxl>=3.1.5

# PySerial - Serial Port Communication
# Required for R307 fingerprint sensor communication via USB
# NOTE: Not needed with ESP32 (sensor connects to ESP32, not PC)
pyserial>=3.5

# Firebase Admin SDK - Firebase Integration
# Required for connecting Django to Firebase Realtime Database
firebase-admin>=6.0.0

# Requests - HTTP Library
# Used for making HTTP requests to Firebase REST API
requests>=2.31.0

# ============================================================
# INSTALLATION INSTRUCTIONS
# ============================================================
# 1. Create virtual environment (recommended):
#    python -m venv venv
#
# 2. Activate virtual environment:
#    - macOS/Linux: source venv/bin/activate
#    - Windows: venv\Scripts\activate
#
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# 4. Run migrations:
#    python manage.py makemigrations
#    python manage.py migrate
#
# 5. Create admin user:
#    python manage.py createsuperuser
#
# 6. Run development server:
#    python manage.py runserver
# ============================================================

